% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weed_Indices_Functions.R
\name{FDR}
\alias{FDR}
\title{Fresh weight to dry weight ratio}
\usage{
FDR(FW, DW)
}
\arguments{
\item{FW}{Fresh weight of weed samples}

\item{DW}{Dry weight of weed samples}
}
\value{
Fresh weight to dry weight ratio
}
\description{
Calculates the Fresh weight to dry weight ratio
}
\examples{
temp1 <- FDR(FW=10,DW=5);
}
