% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weed_Indices_Functions.R
\name{Herbicide_efficiency_index}
\alias{Herbicide_efficiency_index}
\title{Herbicide efficiency index}
\usage{
Herbicide_efficiency_index(YT, YC, WT, WC)
}
\arguments{
\item{YT}{Yield in treated plot}

\item{YC}{Yield in control plot}

\item{WT}{Weed dry weight in treated plot}

\item{WC}{Weed dry weight in control plot}
}
\value{
Herbicide efficiency index
}
\description{
Calculates the Herbicide efficiency index
}
\examples{
temp1 <- Herbicide_efficiency_index(WT=c(15,10,12,8,16),WC=30,YC =3.9,YT =c(5,4.8,4.5,5.5,4.2));
}
