% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weed_Indices_Functions.R
\name{IWMI}
\alias{IWMI}
\title{Integrated weed management index}
\usage{
IWMI(YT, YC, WT, WC)
}
\arguments{
\item{YT}{Yield in treated plot}

\item{YC}{Yield in control plot}

\item{WT}{Weed dry weight in treated plot}

\item{WC}{Weed dry weight in control plot}
}
\value{
Integrated weed management index
}
\description{
Calculates the Integrated weed management index
}
\examples{
temp1 <- IWMI(WT=c(15,10,12,8,16),WC=30,YC =3.9,YT =c(5,4.8,4.5,5.5,4.2));
}
