% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weed_Indices_Functions.R
\name{RDW}
\alias{RDW}
\title{Relative dry weight}
\usage{
RDW(DW)
}
\arguments{
\item{DW}{Dry weight of a weed species per unit area}
}
\value{
Relative dry weight
}
\description{
Calculates the Relative dry weight
}
\examples{
temp1 <- RDW(DW=c(5,4.8,4.5,5.5,4.2));
}
