% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weed_Indices_Functions.R
\name{Weed_control_index}
\alias{Weed_control_index}
\title{Weed control index}
\usage{
Weed_control_index(WDCP, WDTP)
}
\arguments{
\item{WDCP}{Weed dry weight in control plot}

\item{WDTP}{Weed dry weight in treated plot}
}
\value{
Weed control index
}
\description{
Calculates the Weed control index
}
\examples{
temp1 <- Weed_control_index (WDCP =30,WDTP =10);
temp2 <- Weed_control_index( WDCP =30,WDTP =c(10,15,12,8,16) );
}
