% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weed_Indices_Functions.R
\name{Weed_index}
\alias{Weed_index}
\title{Weed index}
\usage{
Weed_index(YFP, YTP)
}
\arguments{
\item{YFP}{Yield in weed free plot}

\item{YTP}{Yield in treated plot}
}
\value{
Weed index
}
\description{
Calculates the Weed index
}
\examples{
temp1 <- Weed_index(YFP = 5.8,YTP =5);
temp2 <- Weed_index(YFP = 5.8,YTP =c(5,4.8,4.5,5.5,4.2) );
}
