% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weed_Indices_Functions.R
\name{Weed_persistence_index}
\alias{Weed_persistence_index}
\title{Weed persistence index}
\usage{
Weed_persistence_index(WT, WC, WPC, WPT)
}
\arguments{
\item{WT}{Weed dry weight in treated plot}

\item{WC}{Weed dry weight in control plot}

\item{WPC}{Weed density in control plot}

\item{WPT}{Weed density in treated plot}
}
\value{
Weed persistence index
}
\description{
Calculates the Weed persistence index
}
\examples{
temp1 <- Weed_persistence_index(WT=c(15,10,12,8,16),WC=30,WPC =80,WPT =c(30,40,25,15,45));
}
