% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weed_Indices_Functions.R
\name{WSE}
\alias{WSE}
\title{Weed smothering efficiency}
\usage{
WSE(SC, IC)
}
\arguments{
\item{SC}{Mean dry weight of weeds in sole crop}

\item{IC}{Mean dry weight of weeds in intercrop}
}
\value{
Weed smothering efficiency
}
\description{
Calculates the Weed smothering efficiency
}
\examples{
temp1 <- WSE(SC=10,IC=5);
}
